Fix possible deadlock in AliasBiMap/AvailableCache
calling `refreshAlias` while holding the mapLock could trigger an async
reload from mapCache.get that could end up being a blocking reload if
the value was quasi concurrently evicted (? - little unclear on how it
ended up blocking, but there is a size based eviction policy, and that
could probably do the trick).

the problem with such a blocking reload is that the reload requires the
mapLock, and still occurs on a different thread. Thus a deadlock where
the task thread waits forever for the mapLock, and the mapLock holder
waits forever for the task to complete.