Fix rebalance-by-host bug
Rebalance by host is not working in our cluster. Because our cluster has multiple minion types per minion, hostToBalance.getMinionTypes() is returning a CSV. Unfortunately hostManager.listHostStatus({CSV}) does not work, since it expects a single value to send to minion.hasType().

Additionally, I believe limiting the hosts sent to getAssignmentsToBalanceHost() to be incorrect. There are likely to be many tasks that don't care if they are moved to a different minion type.

Finally, I believe that all minionType checks are made correctly in the branches below getAssignmentsToBalanceHost(). I followed all the move-things branches in that method, and there was an appropriate minionType check made for each task/job move between hosts.

Refs T63793