addresses replicating-kicking race condition in minion
Currently, failing a minion triggers a large number of replicating tasks, many which become errored due to failed kicks.
I believe this is due to a race condition where the kick message is sent to spawn while the task is idle, then a minion
failure causes spawn to send a task-replicate message to the same minion before the kick message in processesd, the minion
handles the replicate message, finally the minion tries to handle the kick message but fails because the task isnt idle.
This change essentially demotes the non-idle failure to a warning, similar to the minions response when it tries to kick
a task it doesnt have run-capacity for.